Move production check temporal logic
