+++ front/start_worker.ts
@@ -3,7 +3,7 @@ import { setupGlobalErrorHandler } from "@dust-tt/types";
 import { runPostUpsertHooksWorker } from "@app/documents_post_process_hooks/temporal/worker";
 import logger from "@app/logger/logger";
 import { runPokeWorker } from "@app/poke/temporal/worker";
-import { runProductionChecksWorker } from "@app/production_checks/temporal/worker";
+import { runProductionChecksWorker } from "@app/temporal/production_checks/worker";
 import { runScrubWorkspaceQueueWorker } from "@app/temporal/scrub_workspace/worker";
 import { runUpsertQueueWorker } from "@app/temporal/upsert_queue/worker";
 import { runUpdateWorkspaceUsageWorker } from "@app/temporal/usage_queue/worker";
