+++ front/temporal/production_checks/admin/cli.ts
@@ -1,6 +1,6 @@
 import parseArgs from "minimist";
 
-import { launchProductionChecksWorkflow } from "@app/production_checks/temporal/client";
+import { launchProductionChecksWorkflow } from "@app/temporal/production_checks/client";
 
 const main = async () => {
   const argv = parseArgs(process.argv.slice(2));
