+++ front/temporal/production_checks/checks/scrub_deleted_core_document_versions.ts
